+++ android/app/src/androidTest/java/com/thebluefolderproject/leafbyte/compose/AbstractComposeTests.kt
@@ -9,11 +9,11 @@ import androidx.compose.ui.test.ExperimentalTestApi
 import com.thebluefolderproject.leafbyte.ComposeTestFailureException
 import com.thebluefolderproject.leafbyte.TestClock
 import com.thebluefolderproject.leafbyte.activity.LeafByteNavigation
-import com.thebluefolderproject.leafbyte.fragment.DataStoreBackedSettings
-import com.thebluefolderproject.leafbyte.fragment.Settings
-import com.thebluefolderproject.leafbyte.fragment.clearSettingsStore
+import com.thebluefolderproject.leafbyte.google.signin.GoogleSignInManager
 import com.thebluefolderproject.leafbyte.initializeLogInterception
-import com.thebluefolderproject.leafbyte.utils.GoogleSignInManager
+import com.thebluefolderproject.leafbyte.settings.DataStoreBackedSettings
+import com.thebluefolderproject.leafbyte.settings.Settings
+import com.thebluefolderproject.leafbyte.settings.clearSettingsStore
 import com.thebluefolderproject.leafbyte.utils.log
 import de.mannodermaus.junit5.compose.ComposeContext
 import de.mannodermaus.junit5.compose.createComposeExtension
