+++ android/app/src/main/java/com/thebluefolderproject/leafbyte/LeafByteActivity.kt
@@ -14,9 +14,9 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextAlign
 import com.thebluefolderproject.leafbyte.activity.LeafByteNavigation
 import com.thebluefolderproject.leafbyte.activity.ui.theme.LeafByteTheme
+import com.thebluefolderproject.leafbyte.google.signin.isGoogleSignInConfigured
 import com.thebluefolderproject.leafbyte.utils.Text
 import com.thebluefolderproject.leafbyte.utils.getCameraPhotoUri
-import com.thebluefolderproject.leafbyte.utils.isGoogleSignInConfigured
 import com.thebluefolderproject.leafbyte.utils.log
 import com.thebluefolderproject.leafbyte.utils.logError
 import org.opencv.android.OpenCVLoader
