+++ android/app/src/main/java/com/thebluefolderproject/leafbyte/settings/Settings.kt
@@ -2,8 +2,9 @@
  * Copyright © 2024 Abigail Getman-Pickering. All rights reserved.
  */
 
-package com.thebluefolderproject.leafbyte.fragment
+package com.thebluefolderproject.leafbyte.settings
 
+import com.thebluefolderproject.leafbyte.settings.SaveLocation
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.coroutines.flow.Flow
 import net.openid.appauth.AuthState
